Renames ai_synthesis to ai_api_synthesis
