+++ qiskit_ibm_transpiler/wrappers/__init__.py
@@ -33,11 +33,7 @@ Classes
 """
 
 from .ai_routing import AIRoutingAPI
-from .ai_synthesis import (
-    AICliffordAPI,
-    AILinearFunctionAPI,
-    AIPermutationAPI,
-    AIPauliNetworkAPI,
-)
+from .ai_local_routing import AILocalRouting
+from .ai_api_synthesis import AICliffordAPI, AILinearFunctionAPI, AIPermutationAPI
 from .base import BackendTaskError, QiskitTranspilerService
 from .transpile import TranspileAPI, _get_circuit_from_result
